chore(ci): update device configurations and streamline test setup
- Updated CUDA hardware configuration in `ci_mirror_hardwares.yml` to reflect the new device count for `h100_4`.
- Removed hardcoded device settings in `test-merge.yml` and `test-ready.yml`, allowing for dynamic device assignment.
- Introduced a new `run_runner` function in `conftest.py` to handle E2E runner commands with signal forwarding for better process management.
- Refactored tests in `test_deepseek_v2_lite.py` and `test_async_cam_npu.py` to utilize the new `run_runner` function for improved clarity and maintainability.
