+++ .buildkite/cuda/test-ready.yml
@@ -17,6 +17,5 @@ steps:
         commands:
           - "pip install 'lm_eval[api]==0.4.12'"
           - "export AFD_E2E_BACKEND=gpu"
-          - "export AFD_E2E_DEVICES=0,1,2,3"
           - "pytest -sv tests/e2e/models/deepseek_v2_lite/test_deepseek_v2_lite.py"
         mirror_hardwares: h100_4
