+++ tests/e2e/README.md
@@ -8,9 +8,14 @@ The default gate runs four scenarios:
 - `afd-graph`
 - `afd-graph-dbo`
 
-Each scenario evaluates the first 7 GSM8K samples. AFD uses three devices:
-the first two for Attention and the third for FFN. Tests run sequentially and
-must not skip. Every GSM8K evaluation uses 8 few-shot examples.
+Each scenario evaluates the first 7 GSM8K samples. If `AFD_E2E_DEVICES` is set,
+that value is used as-is; otherwise the defaults are:
+
+- `0,1,2` for the default scenarios (Attention on the first two, FFN on the third)
+- `0,1,2,3` for `afd-graph-dbo-2a2f`
+
+Tests run sequentially and must not skip. Every GSM8K evaluation uses 8
+few-shot examples.
 
 See the [E2E testing design](../../docs/design/module/e2e_testing.md) before
 adding a model or case.
@@ -30,7 +35,7 @@ GPU:
 ```bash
 export HF_HOME=/path/to/huggingface
 export AFD_E2E_BACKEND=gpu
-export AFD_E2E_DEVICES=0,1,2
+# Optional: export AFD_E2E_DEVICES=0,1,2
 # Optional if the model is already local:
 # export AFD_GPU_E2E_MODEL=/path/to/DeepSeek-V2-Lite
 ```
@@ -40,7 +45,7 @@ NPU:
 ```bash
 export HF_HOME=/path/to/huggingface
 export AFD_E2E_BACKEND=npu
-export AFD_E2E_DEVICES=0,1,2
+# Optional: export AFD_E2E_DEVICES=0,1,2
 # Optional if the model is already local:
 # export AFD_NPU_E2E_MODEL=/path/to/DeepSeek-V2-Lite
 ```
@@ -59,11 +64,11 @@ Success means 4 passed and 0 skipped.
 
 ### Graph + DBO 2A2F
 
-This separate GPU/NPU case uses four devices: the first two for Attention
-DP=2/TP=1 and the last two for FFN DP=2/TP=1/EP=2. It runs GSM8K-7.
+This separate GPU/NPU case defaults to `0,1,2,3` (or uses `AFD_E2E_DEVICES`
+when set): the first two for Attention DP=2/TP=1 and the last two for FFN
+DP=2/TP=1/EP=2. It runs GSM8K-7.
 
 ```bash
-export AFD_E2E_DEVICES=0,1,2,3
 python -m pytest -q -s \
   "tests/e2e/models/deepseek_v2_lite/test_deepseek_v2_lite.py::test_deepseek_v2_lite[afd-graph-dbo-2a2f]"
 ```
@@ -85,20 +90,19 @@ The repository includes the [`run-e2e`](../../.agents/skills/run-e2e/SKILL.md)
 skill. Open the repository in Codex and ask, for example:
 
 ```text
-Use run-e2e to run the GPU E2E tests with devices 0,1,2 and
-HF_HOME /data/huggingface.
+Use run-e2e to run the GPU E2E tests with HF_HOME /data/huggingface.
 ```
 
-For either backend, provide three device IDs and `HF_HOME`. The model path is
-optional when Hugging Face download is available. The skill checks
-prerequisites, runs the same four tests, and reports failures and process
-cleanup.
+Provide `HF_HOME` and `AFD_E2E_BACKEND`. `AFD_E2E_DEVICES` is optional; when
+unset, the test module picks the defaults above. The model path is optional
+when Hugging Face download is available. The skill checks prerequisites, runs
+the same four tests, and reports failures and process cleanup.
 
 ## NPU async CAM smoke test
 
-This separate test uses four NPUs: the first two for Attention TP=2, and the
-last two for FFN DP=2/TP=1/EP=2. It sends one prompt and requests 32 tokens. It
-does not run GSM8K.
+This separate test still reads `AFD_E2E_DEVICES`. It uses four NPUs: the first
+two for Attention TP=2, and the last two for FFN DP=2/TP=1/EP=2. It sends one
+prompt and requests 32 tokens. It does not run GSM8K.
 
 ```bash
 export AFD_E2E_BACKEND=npu
