+++ tests/e2e/models/deepseek_v2_lite/test_deepseek_v2_lite.py
@@ -5,17 +5,13 @@
 from __future__ import annotations
 
 import os
-import signal
-import subprocess
 import sys
-from contextlib import suppress
 from pathlib import Path
 
 import pytest
 
-from tests.conftest import download_dataset, download_model
+from tests.conftest import download_dataset, download_model, run_runner
 
-REPO_ROOT = Path(__file__).resolve().parents[4]
 GSM8K_DATASET_ID = "openai/gsm8k"
 GSM8K_DATASET_CONFIG = "main"
 DEEPSEEK_V2_LITE_REPO_ID = "deepseek-ai/DeepSeek-V2-Lite"
@@ -26,8 +22,6 @@ SCENARIOS = (
     "afd-graph-dbo",
     "afd-graph-dbo-2a2f",
 )
-# Covers the roughly 64-second nested lm-eval/vLLM cleanup bound with buffer.
-RUNNER_CLEANUP_TIMEOUT_S = 90
 
 
 def _required_env(name: str) -> str:
@@ -35,15 +29,6 @@ def _required_env(name: str) -> str:
     if not value:
         raise RuntimeError(f"{name} must be set")
     return value
-
-
-def _devices(name: str, expected_count: int) -> list[str]:
-    devices = [item.strip() for item in _required_env(name).split(",") if item.strip()]
-    if len(devices) != expected_count:
-        raise RuntimeError(f"{name} must contain exactly {expected_count} devices")
-    if len(devices) != len(set(devices)):
-        raise RuntimeError(f"{name} devices must be unique")
-    return devices
 
 
 def prepare_e2e_assets() -> None:
@@ -81,8 +66,15 @@ def build_runner_command(scenario: str, gsm8k_output_path: Path) -> list[str]:
     else:
         raise RuntimeError("AFD_E2E_BACKEND must be 'gpu' or 'npu'")
 
-    expected_device_count = 4 if scenario == "afd-graph-dbo-2a2f" else 3
-    devices = _devices("AFD_E2E_DEVICES", expected_device_count)
+    env_devices = os.environ.get("AFD_E2E_DEVICES")
+    if env_devices:
+        devices = [item.strip() for item in env_devices.split(",") if item.strip()]
+    else:
+        devices = (
+            ["0", "1", "2", "3"]
+            if scenario == "afd-graph-dbo-2a2f"
+            else ["0", "1", "2"]
+        )
     attention_devices = devices[:2]
     ffn_devices = devices[2:]
     if scenario == "baseline-graph":
@@ -114,59 +106,6 @@ def build_runner_command(scenario: str, gsm8k_output_path: Path) -> list[str]:
     return command
 
 
-def _run_runner(command: list[str], env: dict[str, str] | None = None) -> None:
-    handled_signals = (signal.SIGTERM, signal.SIGINT)
-    previous_handlers = {signum: signal.getsignal(signum) for signum in handled_signals}
-    process: subprocess.Popen | None = None
-    received_signal: int | None = None
-    forwarded = False
-
-    def forward_received_signal() -> None:
-        nonlocal forwarded
-        if process is None or received_signal is None or forwarded:
-            return
-        forwarded = True
-        with suppress(ProcessLookupError):
-            os.killpg(process.pid, received_signal)
-        raise SystemExit(128 + received_signal)
-
-    def forward_cancellation(signum, _frame) -> None:
-        nonlocal received_signal
-        if received_signal is not None:
-            return
-        received_signal = signum
-        forward_received_signal()
-
-    for signum in handled_signals:
-        signal.signal(signum, forward_cancellation)
-
-    try:
-        process = subprocess.Popen(
-            command,
-            cwd=REPO_ROOT,
-            env=env,
-            start_new_session=True,
-        )
-        forward_received_signal()
-        returncode = process.wait()
-        if returncode != 0:
-            raise subprocess.CalledProcessError(returncode, command)
-    finally:
-        try:
-            if process is not None and process.poll() is None:
-                try:
-                    process.wait(timeout=RUNNER_CLEANUP_TIMEOUT_S)
-                except subprocess.TimeoutExpired:
-                    try:
-                        with suppress(ProcessLookupError):
-                            os.killpg(process.pid, signal.SIGKILL)
-                    finally:
-                        process.wait()
-        finally:
-            for signum, previous_handler in previous_handlers.items():
-                signal.signal(signum, previous_handler)
-
-
 @pytest.fixture(scope="module", autouse=True)
 def _prepare_e2e_assets() -> None:
     """Prepare shared E2E assets once for this test module."""
@@ -177,4 +116,4 @@ def _prepare_e2e_assets() -> None:
 @pytest.mark.parametrize("scenario", SCENARIOS, ids=SCENARIOS)
 def test_deepseek_v2_lite(scenario: str, tmp_path: Path) -> None:
     command = build_runner_command(scenario, tmp_path / scenario)
-    _run_runner(command)
+    run_runner(command)
